Create a new container, add content, adjust position
Add a logo.

Create a container for the page title and logo.

Create a 'p' tag for credits.

Add a 'relative position' to the 'first-container' and
'second-container', so we can pass an 'absolute posi-
tion' to the 'logo-container' and move it to the in-
tended place.

We now have two images, therefore working with a vague
'img' tag may create problems. Address the 'background-
image' and 'logo' by their classes to avoid issues.