Style and adjust the forms
Remove the 'form' and it's content from the div
named 'theForms'. If we want to change how the
labels and inputs are distributed within a container,
we simply target the 'form' tag. There is no need to
make it more complex.

Style the forms: Change it's margin, padding, position,
font style, border and color.