Add rules to validate the forms
Add rules that determine what data users can
enter into an input.

Style the forms: if the user enter the right
information into an input, it shows a green border.

Passwords are treated differently. The default
behavior is to highlight an error by display a red
border. Only after reaching the right answer, the
green border is displayed.