Elaborate the script to check the passwords
Stablish a link between the index file and the
JavaScript file.

Create a function to check that the password fi-
elds match each other.

Add a 'keyup' event in the method 'addEventListener()'
to trigger the function.

If the passwords do not match, the user will see the message '☒ Passwords do not match', otherwise they see '🗹 Password Matched'.

Disable the button if the passwords do not match.

Style the 'log in' text.